+++ qcrbox/qcrbox/registry/server/database/database.py
@@ -1,5 +1,6 @@
 import os
 import sqlite3
+from typing import Optional
 
 from loguru import logger
 from sqlalchemy import event
@@ -7,12 +8,10 @@ from sqlalchemy.engine import Engine
 from sqlalchemy.exc import IntegrityError
 from sqlmodel import create_engine, Session, select
 from qcrbox.common.msg_specs.sql_models import QCrBoxBaseSQLModel, KeywordDB
+from qcrbox.common import sql_models
 
 connect_args = {"check_same_thread": False}
 registry_db_dir = os.environ.get("QCRBOX_REGISTRY_DB_DIR", "/mnt/qcrbox/qcrbox_registry_data/")
-if not os.path.exists(registry_db_dir):
-    logger.debug(f"Creating directory for qcrbox registry database: {registry_db_dir!r}")
-    os.makedirs(registry_db_dir)
 
 sqlite_file_name = os.path.join(registry_db_dir, "qcrbox_registry_database.db")
 sqlite_url = f"sqlite:///{sqlite_file_name}"
@@ -65,11 +64,33 @@ def retrieve_application(name, version):
         return result
 
 
-def retrieve_command(name, parameters, application_id):
+def retrieve_command(name: str, parameters, application_id: int):
     from qcrbox.common.msg_specs.sql_models import QCrBoxCommandDB as cls
 
     with Session(engine) as session:
         result = session.exec(
             select(cls).where(cls.name == name, cls.parameters == parameters, cls.application_id == application_id)
         ).one()
+        return result
+
+
+def retrieve_container(container_id: int):
+    from qcrbox.common.msg_specs.sql_models import QCrBoxContainerDB as cls
+
+    with Session(engine) as session:
+        result = session.exec(select(cls).where(cls.id == container_id)).one()
+        return result
+
+
+def retrieve_containers(application_id: Optional[int] = None, command_id: Optional[int] = None):
+    if command_id is not None:
+        raise NotImplementedError("Filtering by command_id is not implemented yet.")
+
+    with Session(engine) as session:
+        stmt = select(sql_models.QCrBoxContainerDB)
+        if application_id is not None:
+            stmt = stmt.where(sql_models.QCrBoxContainerDB.application_id == application_id)
+        # if command_id is not None:
+        #     stmt = stmt.where(command_id in sql_models.QCrBoxContainerDB.commands)
+        result = session.exec(stmt).all()
         return result